requirements.txt
matplotlib
pandas
sklearn
seaborn
pick
pyinputplus
mysql-connector-python
click
fuzzyfinder
prompt-toolkit
terminaltables
